% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_num_factors.R
\name{EstNumBaiNg}
\alias{EstNumBaiNg}
\title{Estimate number of factors according to Bai and Ng (2002).}
\usage{
EstNumBaiNg(Y, r.max = 10, center = TRUE)
}
\arguments{
\item{Y}{Data matrix.}

\item{r.max}{Maximum number of factors (r.max = 10 by default).}

\item{center}{Either TRUE (default) indicating that data matrix must be
centered first, or FALSE.}
}
\value{
The estimated number of factors (between 1 and r.max).
}
\description{
Estimate number of factors according to Bai and Ng (2002).
}
\references{
Bai, Jushan and Serena Ng (2002). “Determining the number of factors in
approximate factor models”. In: Econometrica 70.1, pp. 191–221.
}
