% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_num_factors.R
\name{EstNumKraus}
\alias{EstNumKraus}
\title{Estimate number of factors similar to Kraus (2015).}
\usage{
EstNumKraus(O.set, Y.obs, r.max = 10)
}
\arguments{
\item{O.set}{Set on which functions are observable.}

\item{Y.obs}{Imperfectly observed data matrix.}

\item{r.max}{Maximum number of factors (r.max = 10 by default).}
}
\value{
The estimated number of factors (between 1 and r.max).
}
\description{
Estimate number of factors similar to Kraus (2015).
}
\references{
Kraus, David (2015). “Components and completion of partially observed
functional data”. In: J. R. Stat. Soc. Ser. B. Stat. Methodol. 77.4,
pp. 777–801.
}
