% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_num_factors.R
\name{EstNumOnatski}
\alias{EstNumOnatski}
\title{Estimate number of factors according to Onatski (2010).}
\usage{
EstNumOnatski(Y, r.max = 10, center = TRUE, niter = 10)
}
\arguments{
\item{Y}{Data matrix.}

\item{r.max}{Maximum number of factors (r.max = 10 by default).}

\item{center}{Either TRUE (default) indicating that data matrix must be
centered first, or FALSE.}

\item{niter}{Maximum number of iterations (niter = 10 by default).}
}
\value{
The estimated number of factors (between 1 and r.max).
}
\description{
Estimate number of factors according to Onatski (2010).
}
\references{
Onatski, Alexei (2010). “Determining the number of factors from empirical
distribution of eigenvalues”. In: The Review of Economics and Statistics
92.4, pp. 1004–1016.
}
