% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obs.R
\name{GenObs}
\alias{GenObs}
\title{Generate functional data}
\usage{
GenObs(T = 100, N = 51, r = 10, complete = TRUE, type.miss = "A")
}
\arguments{
\item{T}{Sample size (T = 100 by default).}

\item{N}{Number of equispaced grid points (N = 51 by default).}

\item{r}{Number of global factors (r = 10 by default); must be an even number.}

\item{complete}{Either TRUE (default) indicating that functions are
observed on the whole domain, or FALSE otherwise.}

\item{type.miss}{Type of missingness mechanism; see paper.}
}
\value{
A list containing the true data matrix 'X', the noisy data matrix 'Y' and
the imperfectly observed data matrix 'Y.obs' with NA's.
}
\description{
Generate functional data
}
\examples{
set.seed(1)
data <- GenObs(T = 100, N = 51, r = 10, complete = FALSE, type.miss = "A")
matplot(t(data$Y.obs), type = "l")
}
