% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_functions.R
\name{PredBand}
\alias{PredBand}
\title{Construct simultaneous prediction bands}
\usage{
PredBand(Y0.c, Y1.c = NULL, O.set, p = 0.95, r.max = 20, ro)
}
\arguments{
\item{Y0.c}{Data matrix for target.}

\item{Y1.c}{Data matrix for covariate (NULL for univariate case).}

\item{O.set}{Set on which functions are observable.}

\item{p}{Coverage probability for prediction interval (p = 0.95 by
default).}

\item{r.max}{Maximum number of factors (r.max = 20 by default).}

\item{ro}{Number of factors on O.}
}
\value{
A list with the quantile 'q.hat' and shape 'Z.var.sqrt' for constructing
a simultaneous prediction band.
}
\description{
Construct simultaneous prediction bands
}
