% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_functions.R
\name{ReconstFD}
\alias{ReconstFD}
\title{Reconstruct imperfectly observed functional data}
\usage{
ReconstFD(
  Y.obs,
  T.set = 1:dim(Y.obs)[1],
  method = "Kraus",
  pred.band = FALSE,
  p = 0.95,
  r = NULL,
  r.max = 10
)
}
\arguments{
\item{Y.obs}{Data matrix.}

\item{T.set}{Vector indicating functions to be reconstructed. By default,
all functions are reconstructed.}

\item{method}{Method for determining O-specific number of factors. Must be
either 'BaiNg', 'Kraus' (default), 'Onatski' or 'rGiven' (for manual choice).}

\item{pred.band}{Either TRUE or FALSE (default) indicating whether
prediction bands should be returned.}

\item{p}{Coverage probability for prediction intervals (p = 0.95 by
default).}

\item{r}{Manual choice of O-specific number of factors (only if method =
"rGiven").}

\item{r.max}{Maximum number of factors.}
}
\value{
A list with the reconstructed data matrix 'X.hat', the prediction bands
'U.hat' and 'L.hat' (only if pred.band = TRUE) as well as a vector
containing estimated numbers of factors 'r.hat'.
}
\description{
Reconstruct imperfectly observed functional data
}
\examples{
set.seed(1)
data_com <- GenObs(T = 100, N = 51, r = 10, complete = TRUE, type.miss = "A")
data_inc <- GenObs(T = 1, N = 51, r = 10, complete = FALSE, type.miss = "A")
data <- list(X = rbind(data_com$X, data_inc$X),
             Y = rbind(data_com$Y, data_inc$Y),
             Y.obs = rbind(data_com$Y.obs, data_inc$Y.obs))

# Reconstruct functions
reconst <- ReconstFD(data$Y.obs, pred.band = TRUE, method = "Onatski")

# Plot reconstructions
t.idx <- which(rowSums(is.na(data$Y.obs)) > 0)[1]
lim <- range(data$X[t.idx], data$Y.obs[t.idx,], reconst$U.hat[t.idx,],
             reconst$L.hat[t.idx,], na.rm = TRUE)
plot(data$X[t.idx,], type = "l", lty = 3, ylim = lim)
points(data$Y.obs[t.idx,])
lines(reconst$X.hat[t.idx,], lwd = 2)
lines(reconst$U.hat[t.idx,], lty = 2)
lines(reconst$L.hat[t.idx,], lty = 2)
}
\references{
Bai, Jushan and Serena Ng (2002). “Determining the number of factors in
approximate factor models”. In: Econometrica 70.1, pp. 191–221.

Kraus, David (2015). “Components and completion of partially observed
functional data”. In: J. R. Stat. Soc. Ser. B. Stat. Methodol. 77.4,
pp. 777–801.

Onatski, Alexei (2010). “Determining the number of factors from empirical
distribution of eigenvalues”. In: The Review of Economics and Statistics
92.4, pp. 1004–1016.
}
