% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obs.R
\name{GenObs}
\alias{GenObs}
\title{Generate functional data}
\usage{
GenObs(
  T = 100,
  n = 51,
  r = 50,
  type.miss = "A",
  ev = "exp",
  eps.sd = 0.1,
  complete = FALSE
)
}
\arguments{
\item{T}{Sample size (T = 100 by default).}

\item{n}{Number of equispaced grid points (n = 51 by default).}

\item{r}{Number of factors (r = 50 by default); must be an even number.}

\item{type.miss}{Type of missingness mechanism.}

\item{ev}{Decay of eigenvalues; either "exp" (default) or polynomial else.}

\item{eps.sd}{Standard deviation of measurement errors.}

\item{complete}{If complete = TRUE (default), functions are completely
observable.}
}
\value{
A list containing the true data matrices 'X0', 'X1', the noisy data
matrices 'Y0', 'Y1', and the imperfectly observed data matrices
'Y0.obs', 'Y1.obs' with NA's.
}
\description{
Generate functional data
}
\examples{
set.seed(123)
data <- GenObs(T = 4, n = 51)
par(mfrow = c(1,2))
matplot(t(data$Y0.obs), type = "l")
matplot(t(data$Y1.obs), type = "l")
}
