% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_functions.R
\name{PredBand}
\alias{PredBand}
\title{Construct simultaneous prediction bands}
\usage{
PredBand(Yc, O.set, p = 0.95, r.max = 10, ro)
}
\arguments{
\item{Yc}{Data matrix.}

\item{O.set}{Set on which functions are observable.}

\item{p}{Coverage probability for prediction interval (p = 0.95 by
default).}

\item{r.max}{Maximum number of factors (r.max = 10 by default).}

\item{ro}{Number of factors on O.}
}
\value{
A list with the quantile 'c.hat' and shape 'Z.var.sqrt' for constructing
a simultaneous prediction band.
}
\description{
Construct simultaneous prediction bands
}
